# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Business Management API",
        "description": "Laravel 11 Business Management API with User Model and Sanctum Authentication",
        "contact": {
            "name": "API Support",
            "email": "admin@businessapi.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.yourdomain.com",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/api/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get admin dashboard data",
                "operationId": "a4e953abeda97cdeaa1c2af786aea1f3",
                "responses": {
                    "200": {
                        "description": "Dashboard data retrieved"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/admin/businesses": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create a new business",
                "operationId": "01cbb2c6194deb1932bf9485372f6821",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "business_type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tech Company"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "tech@company.com"
                                    },
                                    "business_type": {
                                        "type": "string",
                                        "example": "Technology"
                                    },
                                    "available_balance": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Business created successfully"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/auth/admin/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Admin login",
                "operationId": "33e473a6771415fc8fe409a086250bd2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "type": "string"
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "abilities": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/auth/business/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Business login",
                "operationId": "05f282b19e160be7e7d8197d287806c7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "business@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "422": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "operationId": "df9bdc6c7a6d9f20b983526b2ec0f698",
                "responses": {
                    "200": {
                        "description": "Logout successful"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get authenticated user",
                "operationId": "b5de204f5e452ef51327c691aa99e77c",
                "responses": {
                    "200": {
                        "description": "User data retrieved"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "description": "Enter your Sanctum authentication token",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}